requirements.txt
skyfield==1.49
sgp4==2.23
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyyaml>=6.0
